// src/types/school.ts
export interface School {
  index: number;
  province_id: number;
  province_name: string;
  city_id: number;
  city_name: string;
  district_id: number;
  district: string;
  npsn: string;
  school_name: string;
  stage: string;
  status: string;
  street_name: string;
  lat: number;
  long: number;
}